// cargo-ship-frontend/src/services/api/models/PaginatedResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PaginatedResponseDto = {
    /**
     * 数据项列表
     */
    items: Array<any[]>;
    /**
     * 总记录数
     */
    total: number;
    /**
     * 当前页码（从 1 开始）
     */
    page: number;
    /**
     * 每页记录数
     */
    pageSize: number;
    /**
     * 总页数
     */
    totalPages: number;
};